Move logic to constant folding
